# mcp_server/openapi_spec.json
{"openapi":"3.1.0","info":{"title":"AutoPro FastMCP Server","description":"\n# AutoPro FastMCP Server\n\nProduction-ready MCP server for AutoPro Daune project with full orchestrator integration.\n\n## Features\n\n- **Workflow Orchestration**: Execute complex multi-step workflows\n- **Linear Integration**: Create and manage tasks\n- **GitHub Integration**: Create issues and commits\n- **Supabase Integration**: Query and verify database\n- **Testing**: Browser E2E and API tests\n- **GPT Developer Mode**: Special endpoints for GPT integration\n\n## GPT Developer Mode Endpoints\n\nThese endpoints are optimized for GPT assistants:\n\n- `POST /mcp/tools/gpt/orchestrate` - Orchestrate complete workflow\n- `POST /mcp/tools/gpt/create_task` - Create Linear task\n- `POST /mcp/tools/gpt/test` - Run tests (browser or API)\n- `GET /mcp/tools/gpt/status` - Get system status\n\n## Usage\n\n1. Check system health: `GET /health`\n2. Orchestrate workflow: `POST /mcp/tools/gpt/orchestrate`\n3. Track progress via Linear and GitHub\n4. Get comprehensive status: `GET /mcp/tools/gpt/status`\n\n## Authentication\n\nCurrently uses orchestrator's configured credentials.\nNo direct authentication required for mcp_server endpoints.\n","version":"0.2.0","x-gpt-integration":{"enabled":true,"version":"1.0","capabilities":["workflow_orchestration","linear_integration","github_integration","supabase_integration","browser_testing","api_testing"]}},"paths":{"/health":{"get":{"summary":"Health","description":"Health check endpoint","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Health Health Get"}}}}}}},"/mcp/execute":{"post":{"summary":"Mcp Execute","description":"Execute a free-form MCP task\nThis is a high-level endpoint that interprets natural language commands","operationId":"mcp_execute_mcp_execute_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecuteRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecuteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/task/{task_id}/status":{"get":{"summary":"Mcp Task Status","description":"Get task status and result","operationId":"mcp_task_status_mcp_task__task_id__status_get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskStatusResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/workflows/orchestrate":{"post":{"summary":"Mcp Orchestrate Workflow","description":"Orchestrate a complete workflow using the orchestrator\nThis is the main entry point for complex multi-step operations","operationId":"mcp_orchestrate_workflow_mcp_workflows_orchestrate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrchestrateWorkflowRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Mcp Orchestrate Workflow Mcp Workflows Orchestrate Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/linear/task":{"put":{"summary":"Mcp Linear Update Task","description":"Update Linear task via orchestrator","operationId":"mcp_linear_update_task_mcp_tools_linear_task_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LinearUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Mcp Linear Update Task Mcp Tools Linear Task Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Mcp Linear Create Task","description":"Create Linear task via orchestrator","operationId":"mcp_linear_create_task_mcp_tools_linear_task_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LinearTaskRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Mcp Linear Create Task Mcp Tools Linear Task Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/linear/tasks":{"get":{"summary":"Mcp Linear List Tasks","description":"List Linear tasks via orchestrator","operationId":"mcp_linear_list_tasks_mcp_tools_linear_tasks_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Mcp Linear List Tasks Mcp Tools Linear Tasks Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/github/issue":{"post":{"summary":"Mcp Github Create Issue","description":"Create GitHub issue via orchestrator","operationId":"mcp_github_create_issue_mcp_tools_github_issue_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GitHubIssueRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Mcp Github Create Issue Mcp Tools Github Issue Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/github/commit":{"post":{"summary":"Mcp Github Commit","description":"Create Git commit via orchestrator","operationId":"mcp_github_commit_mcp_tools_github_commit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GitHubCommitRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Mcp Github Commit Mcp Tools Github Commit Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/supabase/query":{"post":{"summary":"Mcp Supabase Query","description":"Execute Supabase query via orchestrator","operationId":"mcp_supabase_query_mcp_tools_supabase_query_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SupabaseQueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Mcp Supabase Query Mcp Tools Supabase Query Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/supabase/verify":{"post":{"summary":"Mcp Supabase Verify","description":"Verify database fix via orchestrator","operationId":"mcp_supabase_verify_mcp_tools_supabase_verify_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SupabaseVerifyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Mcp Supabase Verify Mcp Tools Supabase Verify Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/test/browser":{"post":{"summary":"Mcp Browser Test","description":"Execute browser E2E test via orchestrator","operationId":"mcp_browser_test_mcp_tools_test_browser_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrowserTestRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Mcp Browser Test Mcp Tools Test Browser Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/test/api":{"post":{"summary":"Mcp Api Test","description":"Execute API test via orchestrator","operationId":"mcp_api_test_mcp_tools_test_api_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/APITestRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Mcp Api Test Mcp Tools Test Api Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/system/health":{"get":{"summary":"Mcp System Health","description":"Check system health via orchestrator","operationId":"mcp_system_health_mcp_tools_system_health_get","parameters":[{"name":"detailed","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Detailed"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Mcp System Health Mcp Tools System Health Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/tools/gpt/orchestrate":{"post":{"summary":"Gpt Orchestrate","description":"GPT Developer Mode: Orchestrate workflow\nSpecial endpoint for GPT integration with enhanced response formatting\n\n**GPT Note**: This endpoint returns GPT-optimized responses.","operationId":"gpt_orchestrate_mcp_tools_gpt_orchestrate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrchestrateWorkflowRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Gpt Orchestrate Mcp Tools Gpt Orchestrate Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"tags":["GPT Developer Mode"]}},"/mcp/tools/gpt/create_task":{"post":{"summary":"Gpt Create Task","description":"GPT Developer Mode: Create Linear task with GPT-friendly response\n\n**GPT Note**: This endpoint returns GPT-optimized responses.","operationId":"gpt_create_task_mcp_tools_gpt_create_task_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LinearTaskRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Gpt Create Task Mcp Tools Gpt Create Task Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"tags":["GPT Developer Mode"]}},"/mcp/tools/gpt/test":{"post":{"summary":"Gpt Run Test","description":"GPT Developer Mode: Run test (browser or API)\nUnified endpoint for all test types\n\n**GPT Note**: This endpoint returns GPT-optimized responses.","operationId":"gpt_run_test_mcp_tools_gpt_test_post","parameters":[{"name":"test_type","in":"query","required":true,"schema":{"type":"string","title":"Test Type"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","title":"Config"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Gpt Run Test Mcp Tools Gpt Test Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"tags":["GPT Developer Mode"]}},"/mcp/tools/gpt/status":{"get":{"summary":"Gpt System Status","description":"GPT Developer Mode: Get comprehensive system status\nIncludes orchestrator, backend, database, etc.\n\n**GPT Note**: This endpoint returns GPT-optimized responses.","operationId":"gpt_system_status_mcp_tools_gpt_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","title":"Response Gpt System Status Mcp Tools Gpt Status Get"}}}}},"tags":["GPT Developer Mode"]}}},"components":{"schemas":{"APITestRequest":{"properties":{"method":{"type":"string","pattern":"^(GET|POST|PUT|DELETE|PATCH)$","title":"Method"},"url":{"type":"string","title":"Url"},"body":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Body"},"headers":{"anyOf":[{"additionalProperties":{"type":"string"},"type":"object"},{"type":"null"}],"title":"Headers"},"expected_status":{"type":"integer","title":"Expected Status","default":200},"expected_response":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Expected Response"}},"type":"object","required":["method","url"],"title":"APITestRequest"},"BrowserTestRequest":{"properties":{"test_name":{"type":"string","title":"Test Name"},"url":{"type":"string","title":"Url"},"steps":{"items":{"type":"object"},"type":"array","title":"Steps"},"verify_in_db":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Verify In Db"}},"type":"object","required":["test_name","url","steps"],"title":"BrowserTestRequest"},"ExecuteRequest":{"properties":{"task":{"type":"string","title":"Task"},"context":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Context"}},"type":"object","required":["task"],"title":"ExecuteRequest"},"ExecuteResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"},"status":{"type":"string","title":"Status"},"message":{"type":"string","title":"Message"}},"type":"object","required":["task_id","status","message"],"title":"ExecuteResponse"},"GitHubCommitRequest":{"properties":{"message":{"type":"string","title":"Message"},"files":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Files"},"linear_task_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Linear Task Id"},"github_issue_number":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Github Issue Number"}},"type":"object","required":["message"],"title":"GitHubCommitRequest"},"GitHubIssueRequest":{"properties":{"title":{"type":"string","title":"Title"},"body":{"type":"string","title":"Body"},"labels":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Labels"},"assignees":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Assignees"},"linear_task_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Linear Task Id"}},"type":"object","required":["title","body"],"title":"GitHubIssueRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LinearTaskRequest":{"properties":{"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"priority":{"type":"integer","title":"Priority","default":0},"labels":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Labels"},"epic_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Epic Id"},"assignee":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Assignee"}},"type":"object","required":["title"],"title":"LinearTaskRequest"},"LinearUpdateRequest":{"properties":{"task_id":{"type":"string","title":"Task Id"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"}},"type":"object","required":["task_id"],"title":"LinearUpdateRequest"},"OrchestrateWorkflowRequest":{"properties":{"command":{"type":"string","title":"Command"},"context":{"type":"object","title":"Context"},"options":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Options"}},"type":"object","required":["command","context"],"title":"OrchestrateWorkflowRequest"},"SupabaseQueryRequest":{"properties":{"table":{"type":"string","title":"Table"},"operation":{"type":"string","pattern":"^(select|insert|update|delete)$","title":"Operation"},"filters":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Filters"},"data":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Data"},"limit":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Limit"}},"type":"object","required":["table","operation"],"title":"SupabaseQueryRequest"},"SupabaseVerifyRequest":{"properties":{"table":{"type":"string","title":"Table"},"expected":{"type":"object","title":"Expected"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["table","expected"],"title":"SupabaseVerifyRequest"},"TaskStatusResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"},"status":{"type":"string","title":"Status"},"result":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Result"}},"type":"object","required":["task_id","status"],"title":"TaskStatusResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"GPT Developer Mode","description":"Special endpoints optimized for GPT assistants with enhanced response formatting"}]}